Replace digits data for cancer data
